lucky/requirements.txt
django==2.2
django-filter==2.1.0
djangorestframework==3.9.4
django-jsonfield-backport==1.0.3